// objectTypes.d.ts
export type booleanString = 'true' | 'false';
export type timestampNumber = number;
export type urlString = `http://${string}` | `https://${string}`;
// ...
export interface GreenButtonSummaryMeasurement {
    powerOfTenMultiplier?: keyof typeof lookups.powerOfTenMultipliers;
    timeStamp?: timestampNumber;
    uom?: keyof typeof lookups.unitsOfMeasurement;
    value?: number;
    readingTypeRef?: urlString;
}
// ...
export interface GreenButtonTariffRider {
    riderType: string;
    enrollmentStatus: keyof typeof lookups.enrollmentStatuses;
    enrollmentStatus_value?: (typeof lookups.enrollmentStatuses)[keyof typeof lookups.enrollmentStatuses];
    effectiveDate: timestampNumber;
}
export interface GreenButtonPNode {
    apnodeType: keyof typeof lookups.pnodeTypes;
    ref: string;
    startEffectiveDate?: timestampNumber;
    endEffectiveDate?: timestampNumber;
}
export interface GreenButtonAggregateNode {
    anodeType: keyof typeof lookups.anodeTypes;
    ref: string;
    startEffectiveDate?: timestampNumber;
    endEffectiveDate?: timestampNumber;
}
export interface GreenButtonUsagePoint {
    roleFlags?: string;
    ServiceCategory?: {
        kind: keyof typeof lookups.serviceCategoryKinds;
        kind_value: (typeof lookups.serviceCategoryKinds)[keyof typeof lookups.serviceCategoryKinds];
    };
    status?: 0 | 1;
    ServiceDeliveryPoint?: {
        name?: string;
        tariffProfile?: string;
        customerAgreement?: string;
        tariffRiderRefs?: {
            tariffRiderRef?: GreenButtonTariffRider | GreenButtonTariffRider[];
        };
    };
    amiBillingReady?: keyof typeof lookups.amiBillingReadyStatuses;
    checkBilling?: booleanString;
    connectionState?: keyof typeof lookups.connectionStates;
    estimatedLoad?: GreenButtonSummaryMeasurement;
    grounded?: booleanString;
    isSdp?: booleanString;
    isVirtual?: booleanString;
    minimalUsageExpected?: booleanString;
    nominalServiceVoltage?: GreenButtonSummaryMeasurement;
    outageRegion?: string;
    phaseCode?: keyof typeof lookups.phaseCodes;
    ratedCurrent?: GreenButtonSummaryMeasurement;
    ratedPower?: GreenButtonSummaryMeasurement;
    readCycle?: string;
    readRoute?: string;
    serviceDeliveryRemark?: string;
    servicePriority?: string;
    pnodeRefs?: {
        pnodeRef: GreenButtonPNode | GreenButtonPNode[];
    };
    aggregateNodeRefs?: {
        aggregateNodeRef: GreenButtonAggregateNode | GreenButtonAggregateNode[];
    };
}
export interface GreenButtonCostAdditionalDetail {
    itemKind: keyof typeof lookups.costDetailItemKinds;
    itemKind_value?: (typeof lookups.costDetailItemKinds)[keyof typeof lookups.costDetailItemKinds];
    note: string;
    amount?: number;
    rounding?: number;
    dateTime?: timestampNumber;
    measurement?: GreenButtonSummaryMeasurement;
    unitCost?: GreenButtonSummaryMeasurement;
    itemPeriod?: GreenButtonSummaryMeasurement;
}